♻️ Refactor segments to optional catch-all segments
refactor segments to optional catch-all segments for merging folder, shared page
